Make edges API
